requirements.txt
flask
opencv-python
numpy
pycryptodome
